selftests: mptcp: Restore packet capture option in join tests
Git-commit: a673321aa74fc5604643d6a4653684c0bc9fa617
Message-Id: <a673321aa74fc5604643d6a4653684c0bc9fa617.1637073710.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.12-rc5

The join self tests previously used the '-c' command line option to
enable creation of pcap files for the tests that run, but the change to
allow running a subset of the join tests made overlapping use of that
option.

Restore the capture functionality with '-c' and move the syncookie test
option to '-k'.

Fixes: 1002b89f23ea ("selftests: mptcp: add command line arguments for mptcp_join.sh")
Acked-and-tested-by: Geliang Tang <geliangtang@gmail.com>
Co-developed-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.12-rc5-selftests-mptcp-Restore-packet-capture-option-in-joi.patch